refactor: move tests to more appropriate package
